+++ packages/gamut-icons/icon-template.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const { startCase } = require('lodash');
 
-function iconTemplate(api, opts, { jsx /* imports, props, exports */ }) {
+function iconTemplate(api, opts, { jsx }) {
   const template = api.template.smart({ plugins: ['jsx', 'typescript'] });
   const { componentName, filePath } = opts.state;
   const exportName = componentName.replace('Svg', '');
@@ -9,23 +9,16 @@ function iconTemplate(api, opts, { jsx /* imports, props, exports */ }) {
 
   return template.ast`
     import * as React from 'react';
-    import { ClassNames } from '@emotion/react';
-    import { getForwardableProps, GamutIconProps, iconStyles } from '../../props';
+    import { Svg, GamutIconProps } from '../../props';
 
     export const ${exportName} = React.forwardRef<SVGSVGElement, GamutIconProps>(({
       title = "${title}",
       titleId,
-      size,
-      className: cn,
-      ...rest
+      ...props
     },
       svgRef
     ) => {
-      const props = getForwardableProps(rest);
-      return <ClassNames>{({ css, cx, theme }) => {
-        const classNames = cx(cn, css(iconStyles({ ...rest, theme })));
-        return ${jsx}
-      }}</ClassNames>;
+      return ${jsx};
     });
   `;
 }